LayoutTests/fast/shapes/shape-inside/shape-inside-calc-crash.html

<!DOCTYPE html>
<html>
<head>
<!--
This is a regression test for https://bugs.webkit.org/show_bug.cgi?id=121020
-->
<script src="../../../resources/js-test-pre.js"></script>
<script>
window.onload = function() {
    shouldBe("window.getComputedStyle(document.getElementById('shape-inside'))['-webkit-shape-inside']", "'circle(50% at 30% 50%)'");
};
</script>
<style>
#shape-inside {
    -webkit-shape-inside: circle(50% at calc(50% - 20%) 50%);
}
</style>
<body>
  <div>This test should not crash</div>
  <div id="shape-inside">Hello</div>
</body>
</html>
